% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecp.R
\name{ecp}
\alias{ecp}
\title{Emulate the conditional posterior}
\usage{
ecp(
  gamma_train,
  gamma_pred,
  lp_alpha,
  p = 1,
  init = rep(0, p),
  m = 1,
  Fdist = "laplace",
  emulator = "gp_mean",
  em_fit = NULL,
  em_pred = NULL,
  nsamp = 15000,
  nburn = 10000,
  nboot = 5,
  ...
)
}
\arguments{
\item{gamma_train}{samples of \gamma for which the conditional posterior of \alpha is sampled. These samples are used as inputs to the emultor.}

\item{gamma_pred}{samples of \gamma used in phase 2 of the ECP algorithm. These samples define the set of conditional posterior distributions of alpha that are used for the final approximation to the cut posterior}

\item{lp_alpha}{log posterior of \alpha|\gamma,y. The first argument should be \alpha, and additional arguments are given through '...'}

\item{p}{dimension of \alpha}

\item{init}{values of \alpha used to initialize MCMC chains targeting the conditional posterior}

\item{m}{number of samples drawn from each approximate conditional posterior in the final approximation to the cut posterior}

\item{Fdist}{distribution used to estimate the conditional posterior. Default is 'normal', which generalizes to the multivariate normal for p>1}

\item{emulator}{emulation algorithm used for estimating the conditional posterior parameters (gp_mean, gp_sample, custom)}

\item{em_fit}{fit function for emulator only used when emulator='custom'. First input should be X matrix (inputs) and second should be y vector (outputs)}

\item{em_pred}{Prediction function for emulator only used when emulator='custom'. Should return numeric vector of predictions at input matrix X.}

\item{nsamp}{number of MCMC samples drawn from each conditional posterior in phase 1}

\item{nburn}{number of burn in samples for each conditional posterior in phase 1}

\item{nboot}{number of bootstrap samples for parameter estimation}

\item{...}{additional arguments passed to lp_alpha}
}
\value{
Description of the return value.
}
\description{
Implements the ECP algorithm, which emulates the conditional posterior distribution to approximate the cut posterior distribution
}
\examples{
An example problem can be found in examples/diamond_in_box
}
