% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecp.R
\name{ecp2}
\alias{ecp2}
\title{Emulate the conditional posterior}
\usage{
ecp2(
  gamma_train,
  gamma_pred,
  alpha_train_list,
  p = ncol(alpha_train_list[[1]]),
  init = rep(0, p),
  m = 1,
  alpha_transform = "identity",
  Fdist = "normal",
  emulator = "gp_mean",
  em_fit = NULL,
  em_pred = NULL,
  nsamp = 15000,
  nburn = 10000,
  nboot = 5,
  ...
)
}
\arguments{
\item{gamma_train}{samples of \gamma for which the conditional posterior of \alpha is sampled. These samples are used as inputs to the emultor.}

\item{gamma_pred}{samples of \gamma used in phase 2 of the ECP algorithm. These samples define the set of conditional posterior distributions of alpha that are used for the final approximation to the cut posterior}

\item{alpha_train_list}{conditional samples \alpha|\gamma for each \gamma in gamma_train}

\item{p}{dimension of \alpha}

\item{init}{starting values for Fdist parameters}

\item{m}{number of samples drawn from each approximate conditional posterior in the final approximation to the cut posterior}

\item{alpha_transform}{transformation applied to \alpha prior to fitting Fdist. If for example \alpha \in [0,1], 'logit' or 'probit' would be appropriate}

\item{Fdist}{distributional assumption for \alpha|\gamma}

\item{emulator}{emulation algorithm used for predicting the conditional posterior parameters at gamma_pred}

\item{em_fit}{fit function for emulator only used when emulator='custom'. First input should be X matrix (inputs) and second should be y vector (outputs)}

\item{em_pred}{Prediction function for emulator only used when emulator='custom'. Should return numeric vector of predictions at input matrix X.}
}
\value{
Description of the return value.
}
\description{
Implements the ECP algorithm with provided samples from the conditional posterior
}
